+++ src/components/ArtworkPage/ArtworkSearchBar.tsx
