+++ src/components/ArtworkPage/FollowedArtistsSection.tsx
